# pkg/apiclient/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: v0.0.0-dev
servers:
- url: http://localhost:3986/
security:
- Bearer: []
paths:
...
  /project-config/prebuild:
    delete:
      description: Delete prebuild
      operationId: DeletePrebuild
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrebuildConfig'
        description: Prebuild
        required: true
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete prebuild
      tags:
      - prebuild
      x-codegen-request-body-name: prebuild
    get:
      description: List prebuilds
      operationId: ListPrebuilds
      parameters:
      - description: Config name
        in: path
        name: configName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/PrebuildDTO'
                type: array
          description: OK
      summary: List prebuilds
      tags:
      - prebuild
    post:
      description: Set prebuild
      operationId: SetPrebuild
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrebuildDTO'
        description: Prebuild
        required: true
      responses:
        "201":
          content: {}
          description: Created
      summary: Set prebuild
      tags:
      - prebuild
      x-codegen-request-body-name: prebuild
  /project-config/prebuild/process-git-event:
    post:
      description: ProcessGitEvent
      operationId: ProcessGitEvent
      requestBody:
        content:
          '*/*':
            schema:
              type: object
        description: Webhook event
        required: true
      responses:
        "200":
          content: {}
          description: OK
      summary: ProcessGitEvent
      tags:
      - prebuild
      x-codegen-request-body-name: workspace
...
components:
  schemas:
...
    PrebuildConfig:
      example:
        commitInterval: 0
        id: id
        branch: branch
        triggerFiles:
        - triggerFiles
        - triggerFiles
      properties:
        branch:
          type: string
        commitInterval:
          type: integer
        id:
          type: string
        triggerFiles:
          items:
            type: string
          type: array
      type: object
    PrebuildDTO:
      example:
        projectConfigName: projectConfigName
        commitInterval: 0
        branch: branch
        triggerFiles:
        - triggerFiles
        - triggerFiles
      properties:
        branch:
          type: string
        commitInterval:
          type: integer
        projectConfigName:
          type: string
        triggerFiles:
          items:
            type: string
          type: array
      type: object
...
  securitySchemes:
    Bearer:
      description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
      in: header
      name: Authorization
      type: apiKey
...